refactor(register): move into pages folder
